Updated DVL demo "robot" to use eNovus mesh and dimensions
Also updated WHOI DVL urdf files to include an "enable_water_tracking"
element in the plugin parameters.  When enabled (1), the DVL plugin
will include water tracking results (based on robot base link motion,
so it won't account for current yet) if the NPS additions to the WHOI
DSL environment are used (it will work fine with the WHOI version).