Added a side-by-side visual with the 3 DVL meshes we've got so far
The model includes the Teledyne WHN, Teledyne Explorer and Nortek
DVL1000 and is launched from the command line with
"roslaunch nps_uw_sensors_gazebo whoi_dvl_static_side_by_side.launch"

Gravity is disabled so that the plugins (WHOI version) will publish
data (even though none of the DVLs is moving) without the DVLs
falling to the ground.  An RVIZ visualization is also launched
that actually provides a better lit view of the DVL meshes.