Made some launch, world, & motion file changes ISO DVL testing
The world file change starts the DVL at a lower depth to improve
visualization (so the ocean surface isn't in the way).  A second
simple motion node was implemented to drive the DVL in a straight
line over a descending octagon pattern (straight line makes the
"correctness" of the DVL water tracking easier to see).