Added point cloud calculation back into NpsGazeboRosImageSonar
The point cloud is only published if there is a subscriber to the ROS
topic, but it is computed if the sensor is active for use in the
sonar calculation as required.  Point cloud results are stored in the
PointCloud2 message and LOS range from the sensor to the point is also
stored in a cv::Mat image (CV_32FC1) as well.